Fix CI: resolve 20 clippy lints + getrandom wasm_js for test target
Clippy fixes (auto-fix + manual):
- Remove unnecessary u32/f64 casts in events.rs, mini_calendar.rs, image_upload.rs
- Replace manual clamp with .clamp() in message_input.rs
- Remove redundant closures in registrations.rs, breadcrumb.rs
- Fix needless borrow in settings.rs
- Replace .clone() with copy in reaction_bar.rs
- Fix manual strip_prefix in mention_text.rs
- Remove useless conversion in image_upload.rs
- Fix let_and_return in virtual_list.rs
- Fix manual_div_ceil in mini_calendar.rs

Test wasm fix:
- Add getrandom 0.3 with wasm_js feature as dev-dependency for
  wasm32 target in nostr-core (proptest → rand 0.9 → getrandom 0.3)

+++ community-forum-rs/crates/forum-client/src/pages/events.rs
@@ -127,16 +127,16 @@ pub fn EventsPage() -> impl IntoView {
         let evts = events.get();
         let now = now_ts;
         let d = js_sys::Date::new_0();
-        let current_month = d.get_month() as u32;
-        let current_year = d.get_full_year() as u32;
+        let current_month = d.get_month();
+        let current_year = d.get_full_year();
 
         evts.iter()
             .filter(|e| e.end_time >= now) // only upcoming
             .filter_map(|e| {
                 let ed = js_sys::Date::new_0();
                 ed.set_time((e.start_time as f64) * 1000.0);
-                let m = ed.get_month() as u32;
-                let y = ed.get_full_year() as u32;
+                let m = ed.get_month();
+                let y = ed.get_full_year();
                 if y == current_year && m == current_month {
                     Some(ed.get_date())
                 } else {
